requirements.txt
matplotlib==3.8.3
pandas==2.2.1
Pillow==10.2.0
seaborn==0.13.2
streamlit==1.32.0
